refactor: Para ficar mais claro o objetivo de cada projeto, renomeada solução, e projetos
