+++ CleanArchitecture.ISP.DeviceSync.Domain/Brands/Entities/Brand.cs
@@ -1,0 +1,6 @@
+﻿namespace CleanArchitecture.ISP.DeviceSync.Domain.Brands.Entities;
+public enum Brand
+{
+    ControlID,
+    Henry
+}
