+++ CleanArchitecture.ISP.DeviceSync.Domain/CleanArchitecture.ISP.DeviceSync.Domain.csproj
