+++ CleanArchitecture.ISP.DeviceSync.Domain/OperationResult/SetDateTime.cs
@@ -1,4 +1,4 @@
-﻿namespace CleanArchitecture.ISP.Domain.OperationResult;
+﻿namespace CleanArchitecture.ISP.DeviceSync.Domain.OperationResult;
 public record SetDateTimeSuccess(string PointClockIp) : Result(true)
 {
     public DateTime SettedDateTime => DateTime.UtcNow;
