+++ CleanArchitecture.ISP.DeviceSync.Domain/PointClocks/Gateway/IControlIDGateway.cs
@@ -1,0 +1,7 @@
+﻿using CleanArchitecture.ISP.DeviceSync.Domain.OperationResult;
+
+namespace CleanArchitecture.ISP.DeviceSync.Domain.PointClocks.Gateway;
+public interface IControlIDGateway
+{
+    Result ShareBiometrics();
+}
