+++ CleanArchitecture.ISP.DeviceSync.Domain/PointClocks/Gateway/IPointClockGenericGateway.cs
@@ -1,0 +1,7 @@
+﻿using CleanArchitecture.ISP.DeviceSync.Domain.OperationResult;
+
+namespace CleanArchitecture.ISP.DeviceSync.Domain.PointClocks.Gateway;
+public interface IPointClockGenericGateway
+{
+    Result SetDateTime();
+}
