+++ CleanArchitecture.ISP.DeviceSync/CleanArchitecture.ISP.DeviceSync.csproj
@@ -8,7 +8,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\CleanArchitecture.ISP.Domain\CleanArchitecture.ISP.Domain.csproj" />
+    <ProjectReference Include="..\CleanArchitecture.ISP.Domain\CleanArchitecture.ISP.DeviceSync.Domain.csproj" />
   </ItemGroup>
 
 </Project>
