--- CleanArchitecture.ISP.Domain/Brands/Entities/Brand.cs
@@ -1,6 +1,0 @@
-﻿namespace CleanArchitecture.ISP.Domain.Brands.Entities;
-public enum Brand
-{
-    ControlID,
-    Henry
-}
