--- CleanArchitecture.ISP.Domain/PointClocks/Gateway/IControlIDGateway.cs
@@ -1,7 +1,0 @@
-﻿using CleanArchitecture.ISP.Domain.OperationResult;
-
-namespace CleanArchitecture.ISP.Domain.PointClocks.Gateway;
-public interface IControlIDGateway
-{
-    Result ShareBiometrics();
-}
