--- CleanArchitecture.ISP.Domain/PointClocks/Gateway/IPointClockGenericGateway.cs
@@ -1,7 +1,0 @@
-﻿using CleanArchitecture.ISP.Domain.OperationResult;
-
-namespace CleanArchitecture.ISP.Domain.PointClocks.Gateway;
-public interface IPointClockGenericGateway
-{
-    Result SetDateTime();
-}
